server/requirements.txt
git+https://github.com/pyeve/eve@0df8e7c6c78455fc56c20a15584afe407642bdba
flask-cors
requests
jinja2
